requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.6
chromadb>=0.4.0
psutil>=5.9.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
rich>=13.0.0
click>=8.1.0